+++ src/ts/index.ts
@@ -1,11 +1,12 @@
 import '../scss/_main.scss';
 import * as ui from './ui';
 import * as loadMap from './loadMap';
-import { sanitizeBeatSaverID, sanitizeURL } from './utils';
+import { sanitizeBeatSaverID, sanitizeURL, isHex } from './utils';
 
 const url = new URL(location.href);
 const link = url.searchParams.get('url');
 const id = url.searchParams.get('id');
+const hash = url.searchParams.get('hash');
 
 ui.init();
 
@@ -13,4 +14,6 @@ if (link) {
     loadMap.downloadFromURL(sanitizeURL(decodeURI(link)));
 } else if (id) {
     loadMap.downloadFromID(sanitizeBeatSaverID(decodeURI(id)));
+} else if (hash && isHex(decodeURI(hash).trim())) {
+    loadMap.downloadFromHash(decodeURI(hash).trim());
 }